+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## **3.1.1**
 ### Fixes
-Add headers to collection requests made using the Connected Systems API. This fixes authentication issues when collection requests are made.
+- Add headers to collection requests made using the Connected Systems API. This fixes authentication issues when collection requests are made.
+- Fixed command and command status collection parsing.
+- Corrected `ObservationFilter` to use accurate default query parameters according to OGC API - Connected Systems specification.
+### Additions
+- Added Jest for unit testing.
+- Added control stream and data stream collection tests.
 
 ## **3.1.0**
 ### Changes